move HasMetadataComparator to exported package
The openshift-model needs to make use of the HasMetadataComparator and therefore
it should reside in an exported package

fixes #2309
