+++ kubernetes-model-generator/kubernetes-model-core/src/main/java/io/fabric8/kubernetes/api/model/HasMetadataComparator.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.fabric8.kubernetes.internal;
+package io.fabric8.kubernetes.api.model;
 
 import io.fabric8.kubernetes.api.model.HasMetadata;
